requirements.txt
asgiref==3.3.4
bleach==3.3.0
Django==3.2.4
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-todo==2.4.9
packaging==20.9
Pillow==8.2.0
pyparsing==2.4.7
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
Unidecode==1.2.0
webencodings==0.5.1
